requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
audioread==2.1.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
decorator==5.1.1
fastdtw==0.3.4
ffmpeg==1.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.28.5
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.38.0
MarkupSafe==2.0.1
matplotlib==3.5.1
noisereduce==2.0.0
numba==0.55.0
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pickle-mixin==1.0.2
Pillow==9.0.0
pooch==1.5.2
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
resampy==0.2.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
SoundFile==0.10.3.post1
threadpoolctl==3.0.0
tqdm==4.63.1
urllib3==1.26.8
Werkzeug==2.0.2
